Change startup location and fix some non-updated code for the cache size change
The reason why the program counter value on startup is being changed to 0x00007FE0 is a little complicated. Basically, it makes it easier to load an auto-generated binary that fills more space in the cache. Because of this change, a binary can be loaded starting at 0x00010000 and so on at consecutive addresses and it will not run out of association-lanes (aka ways) until the entire cache-data is full.

Also I forget to change a few things that should have been changed when I changed the cache size. Those have also been fixed.